Refactor and simplify stylesheets + organize files into multiple directories
