+++ src/Application/Application.css
@@ -1,0 +1,69 @@
+body {
+    margin: 0;
+    display: flex;
+    place-items: center;
+    min-width: 320px;
+    min-height: 100vh;
+}
+
+/* root div tag of project */
+#root {
+    height: 100vh;
+    margin: 0 auto;
+
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+
+    display: flex;
+    flex-direction: column;
+}
+
+/* div containing core content */
+#timeline-wrapper {
+    width: 100vw;
+
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+
+    display: flex;
+}
+
+/* class to hide components */
+.hidden {
+    display:none;
+}
+
+/* Container to fill void space for positioning */
+.fill-container {
+    height: 500px;
+
+    flex: 1;
+    display: flex;
+    justify-content: left;
+    align-items: center;
+}
+
+
+/* General Button style */
+button {
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    background-color: #1a1a1a;
+    cursor: pointer;
+    transition: border-color 0.25s;
+}
+
+button:hover {
+    border-color: #646cff;
+}
+
+button:focus,
+button:focus-visible {
+    outline: 4px auto -webkit-focus-ring-color;
+}
