+++ src/DeleteEntryModal/DeleteEntryModal.css
@@ -1,0 +1,27 @@
+/* DeleteEntry modal styling */
+.modal {
+    background-color: white;
+    color: black;
+    border: 1px solid black;
+    padding: 3%;
+    height: 100%;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 30px;
+
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+}
+
+/* Positioning of Yes and No buttons */
+.modal div {
+    display: flex;
+    justify-content: space-around;
+    gap: 100%;
+}
