+++ src/EntryCard/EntryCardDescription.css
@@ -1,41 +1,3 @@
-.card-container {
-    height: 480px;
-    width: 480px;
-    padding: 10px;
-    border: 3px solid black;
-    background-color: white;
-    color: black;
-
-    display: flex;
-    flex-direction: column;
-}
-
-#hidden {
-    display:none;
-}
-
-/* Image styling */
-#image-container {
-    width: 100%;
-    height: auto;
-    max-height: 50%;
-}
-
-.entry-image {
-    width: 95%;
-    height: auto;
-    max-height: 95%;
-    object-fit: cover;
-    outline: 1px solid black;
-    cursor: pointer;
-}
-
-#change-image {
-    width: 0;
-    height: 0;
-    display: none;
-}
-
 /* Title, Date, Description container styling */
 #description-container {
     width: 100%;
@@ -60,6 +22,7 @@
     gap: 5%;
 }
 
+
 /* Title styling */
 .entry-title {
     width: 90%;
@@ -70,34 +33,23 @@
     color: black;
     border: 0 solid transparent;
 
-    /*font-family: ;*/
     font-size: large;
     font-weight: bold;
     border-radius: 1%;
     text-align: center;
 }
 
-.entry-title:hover {
-    outline: 1px solid black;
-    /*padding: 4px 4px 1px;*/
-}
-
-.entry-title:focus {
-    outline: 1px solid black;
-}
 
 /* Date styling */
 .entry-date {
     width: 90%;
     height: auto;
-    /*padding: 1px;*/
     padding: 1px 5px;
 
     background: transparent;
     color: black;
     border: 0 solid transparent;
 
-    /*font-family: ;*/
     font-size: large;
     font-weight: lighter;
     font-style: italic;
@@ -105,14 +57,6 @@
     text-align: center;
 }
 
-.entry-date:hover {
-    outline: 1px solid black;
-    /*padding: 0 4px;*/
-}
-
-.entry-date:focus {
-    outline: 1px solid black;
-}
 
 /* Description styling */
 .entry-description {
@@ -130,20 +74,16 @@
     text-align: left;
 }
 
-.entry-description:hover {
+
+/* Hover and on Select additional style */
+.entry-title:hover, .entry-title:focus,
+.entry-description:hover, .entry-description:focus,
+.entry-date:hover, .entry-date:focus {
     outline: 1px solid black;
-    /*padding: 4px;*/
 }
 
-.entry-description:focus {
-    outline: 1px solid black;
-}
 
- /*Reduced interaction for EntryCard components */
-.reduced-interaction .entry-image:hover {
-    cursor: default;
-}
-
+/* Reduced interaction for EntryCard components */
 .reduced-interaction .entry-title:focus,
 .reduced-interaction .entry-date:focus,
 .reduced-interaction .entry-description:focus,
@@ -151,16 +91,4 @@
 .reduced-interaction .entry-date:hover,
 .reduced-interaction .entry-description:hover {
     outline: 0 solid black;
-}
-
-/* Sidebar styling */
-.arrow {
-    width: 20px;
-    height: 40px;
-    background-color: #888888;
-    border: 1px solid black;
-}
-
-.arrow:hover {
-    background-color: #7c7c7c;
 }
