+++ src/Header/Header.css
@@ -1,3 +1,4 @@
+/* Header region container */
 #header {
     width: 100vw;
     display: flex;
@@ -6,10 +7,10 @@
     align-items: center;
 }
 
+/* Container for view buttons */
 #header-view-buttons {
     border: 1px solid black;
     padding: 1%;
-    /*margin: 3%;*/
     background-color: #2a2929;
     border-radius: 1%;
 
@@ -17,4 +18,9 @@
     display: flex;
     justify-content: center;
     gap:10px;
+}
+
+/* View button on select additional styling */
+.view-select {
+    outline: 4px auto -webkit-focus-ring-color;
 }
